Abstract out gstreamer code and improve export dialogue
The export code is still very much a placeholder, but it
should allow for creating dynamic pipelines on the fly.

Perhaps we want to expose gstreamer to the user so they
can define their own gstreamer pipelines for unsupported
or more advanced use cases.
