+++ src/export/export-gstreamer.h
@@ -1,0 +1,6 @@
+#include "../model/keyframe-composition.h"
+#include "../model/keyframe-layers.h"
+#include "../keyframe-renderer.h"
+
+void
+export_composition_gstreamer (KeyframeComposition *composition);
