+++ src/keyframe-exporter.c
@@ -1,0 +1,5 @@
+// TODO: Not a gobject
+// Task-based export utilities
+
+// Move some of export dialog's functionality into
+// here and the rest into the export subdirectory
